chore: Move PerperFabricConfig so that it is available in Services/
